+++ R/screen_models.R
@@ -57,7 +57,7 @@
 #' screened.
 #'
 #' @seealso \code{\link{check_sampling}}, \code{\link{check_fit}},
-#' \code{\link{amend}}
+#' \code{\link{amend}}, \code{\link{pull_best}}
 #'
 #' @examples
 #' \dontrun{